--- src/images-watches.js
@@ -1,15 +1,0 @@
-import smartwatch_black from './images/smartwatch/smartwatch_black.png'
-import smartwatch_pink from './images/smartwatch/smartwatch_pink.png'
-import smartwatch_yelow from './images/smartwatch/smartwatch_yellow.png'
-import delivery from './images/icons/Delivery.svg'
-import arrow from './images/icons/Arrow.svg'
-
-const ImagesWatches ={
-    smartwatch_black,
-    smartwatch_pink,
-    smartwatch_yelow,
-    delivery,
-    arrow,
-
-};
-export default ImagesWatches;
